🚚Reorganise functions and TODOs across all files
Move related functions into appropriate date_logic.py and validation.py
files within processes module to reduce size of data_handling.py file.
